Wrapper for credible sets from Multi-Group Single-Trait Fine-Mapping with JAM

Wrapper for credible sets from Multi-Group Single-Trait Fine-Mapping with JAM

MGFMwithJAM(
  gwas.list,
  corX.list,
  Nall,
  save.path,
  cpp = 0.99,
  cred = 0.99,
  maxcv = 1,
  maxcv_stop = 20,
  NCORES = 1,
  jam.nM.iter = 5,
  r2 = 0.8,
  extra.java.arguments = NULL
)

Arguments

gwas.list

List of A data.frame objects, where A is the number of groups; gwas.list[[i]] is a data.frame for group i with 3 columns named: rsID, beta, EAF

corX.list

List of A data.frame objects, where A is the number of groups; corX.list[[i]] is the SNP correlation matrix for group i

Nall

Vector of length A; Nall[i] is the (effective) sample size for group i

save.path

Path to save JAM output files; tmp files and could delete these later e.g. save.path=paste0(DIRout,"/tmpJAM/region1").

cpp

cumulative posterior probability threshold for selecting top models; default 0.99

cred

Level for credible set, default 0.99

maxcv

starting value for maximum number of causal variants

maxcv_stop

maximum value to consider for maximum number of causal variants; maxcv_stop >= maxcv

NCORES

number of cores for parallel computing; recommend NCORES=A, but if on Windows, use NCORES=1

jam.nM.iter

in millions, number of iterations to use in JAM; defailt 5 (5 million)

r2

r.squared threshold for to find tag SNPs to reduce model search space before expanding to all SNPs;

extra.java.arguments

A character string to be passed through to the java command line. E.g. to specify a different temporary directory by passing "-Djava.io.tmpdir=/Temp".

Value

List consisting of two objects: CSsummary = List of one data.frame for each trait; each trait data.frame gives the variants in the multi-group credible set for the trait, the mgMPP, pooled MAF, proportion of studies that contain the variant, names of studies that contain the variant CSdetail = [[1]] a list of multi-group credible sets (variants and their mgMPP), one for each trait; [[2]] details = for each trait, list of all multi-group models and variants and their mgPP, mgMPP

Author

Jenn Asimit