Make two sets of SNP groups using fine-mapping information from all of the traits using two sets of results and maps the names between them

Make two sets of SNP groups using fine-mapping information from all of the traits using two sets of results and maps the names between them

makeSNPgroups2U(
  main.input,
  fm.multi,
  is.snpmat,
  min.mppi = 0.01,
  minsnpmppi = 0.001,
  r2.minmerge = 0.5
)

Arguments

main.input

output from flashfm.input function

fm.multi

output from flashfm function

is.snpmat

logical taking value TRUE when genotype matrix is provided and FALSE when covariance matrix is given

min.mppi

trim snp groups with total MPPI < min.mppi in all diseases; default 0.01

minsnpmppi

only group snps with total MPPI > minsnpmppi; default 0.001

r2.minmerge

merge groups with minimum between-group r2 > r2.minmerge; default 0.5

Value

list of three objects: groups.fm is a list of SNP groups using the single-trait results; groups.flashfm is a list of SNP groups using the flashfm results; group.sizes is a table of SNP group sizes for the two sets of groups